+++ test/filter_test.ts
@@ -1,5 +1,4 @@
-import { test } from "https://deno.land/std@v0.32.0/testing/mod.ts";
-import { assertEquals } from "https://deno.land/std@v0.32.0/testing/asserts.ts"
+import { assertEquals } from "https://deno.land/std@v0.33.0/testing/asserts.ts"
 
 import filter from '../lib/filter.ts';
 
@@ -7,7 +6,7 @@ function isLessThan5(n) {
     return n < 5;
 }
 
-test({
+Deno.test({
     name: "filter",
     fn(): void {
         assertEquals(filter(isLessThan5, [1, 2, 6, 8]), [1, 2]);