+++ test/find_test.ts
@@ -1,5 +1,4 @@
-import { test } from "https://deno.land/std@v0.32.0/testing/mod.ts";
-import { assertEquals } from "https://deno.land/std@v0.32.0/testing/asserts.ts"
+import { assertEquals } from "https://deno.land/std@v0.33.0/testing/asserts.ts"
 
 import find from '../lib/find.ts';
 
@@ -17,7 +16,7 @@ const cars = [
     { make: "Toyota", model: "Supra" },
 ];
 
-test({
+Deno.test({
     name: "find",
     fn(): void {
         assertEquals(find(findToyota, cars), { make: "Toyota", model: "Mirai" });