+++ test/flatten_test.ts
@@ -1,9 +1,8 @@
-import { test } from "https://deno.land/std@v0.32.0/testing/mod.ts";
-import { assertEquals } from "https://deno.land/std@v0.32.0/testing/asserts.ts"
+import { assertEquals } from "https://deno.land/std@v0.33.0/testing/asserts.ts"
 
 import flatten from '../lib/flatten.ts';
 
-test({
+Deno.test({
     name: "flatten",
     fn(): void {
         assertEquals(flatten([[1, 2, 3], [4, 5, 6]]), [1, 2, 3, 4, 5, 6]);