+++ test/identity_test.ts
@@ -1,9 +1,8 @@
-import { test } from "https://deno.land/std@v0.32.0/testing/mod.ts";
-import { assertEquals } from "https://deno.land/std@v0.32.0/testing/asserts.ts"
+import { assertEquals } from "https://deno.land/std@v0.33.0/testing/asserts.ts"
 
 import identity from '../lib/identity.ts';
 
-test({
+Deno.test({
     name: "identity",
     fn(): void {
         assertEquals(identity(1), 1);