+++ test/map_test.ts
@@ -1,5 +1,4 @@
-import { test } from "https://deno.land/std@v0.32.0/testing/mod.ts";
-import { assertEquals } from "https://deno.land/std@v0.32.0/testing/asserts.ts"
+import { assertEquals } from "https://deno.land/std@v0.33.0/testing/asserts.ts"
 
 import map from '../lib/map.ts';
 
@@ -7,7 +6,7 @@ function double(n) {
     return n * 2;
 }
 
-test({
+Deno.test({
     name: "map",
     fn(): void {
         assertEquals(map(double, [1, 2, 3]), [2, 4, 6]);