+++ test/memoize_test.ts
@@ -1,5 +1,4 @@
-import { test } from "https://deno.land/std@v0.32.0/testing/mod.ts";
-import { assertEquals } from "https://deno.land/std@v0.32.0/testing/asserts.ts";
+import { assertEquals } from "https://deno.land/std@v0.33.0/testing/asserts.ts";
 
 import memoize from '../lib/memoize.ts';
 
@@ -7,7 +6,7 @@ function double(n: number) {
     return n * 2;
 }
 
-test({
+Deno.test({
     name: "memoize",
     fn(): void {
         assertEquals(double(0), memoize(double)(0));