+++ test/values_test.ts
@@ -1,9 +1,8 @@
-import { test } from "https://deno.land/std@v0.32.0/testing/mod.ts";
-import { assertEquals } from "https://deno.land/std@v0.32.0/testing/asserts.ts"
+import { assertEquals } from "https://deno.land/std@v0.33.0/testing/asserts.ts"
 
 import values from '../lib/values.ts';
 
-test({
+Deno.test({
     name: "values",
     fn(): void {
         assertEquals(values({ a: 1, b: 2 }), [1, 2]);