+++ lib/either/error.ts
@@ -54,7 +54,7 @@ export default function error<Left, Right>(value: Right): Either<Left, Right> {
         toString() {
             return String(value);
         },
-        [Deno.symbols.customInspect]() {
+        [Deno.customInspect]() {
             return `error(${value})`;
         }
     }