+++ lib/either/json_error.ts
@@ -62,7 +62,7 @@ export default function jsonError<Left, Right>(value: Right): Either<Left, Right
         toString() {
             return String(value);
         },
-        [Deno.symbols.customInspect]() {
+        [Deno.customInspect]() {
             return `jsonError(${value})`;
         }
     }