+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - flatten function flattens by one level instead of Infinity
 - locked all std dependencies (`std/testing`) to `v0.32.0`
 - added better typings for places where types could be ambigous
+- migrated to `deno test`
+- bumped `std` version to `v0.33.0`
 
 ### Removed