+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added better typings for places where types could be ambigous
 - migrated to `deno test`
 - bumped `std` version to `v0.33.0`
+- fixes `prop` failing TS checks when being used
 
 ### Removed